dmarket_scraperRust/requirements.txt
selenium==4.15.2
pyotp==2.9.0
python-dotenv==1.0.0